requirements.txt
mcp>=1.26.0
python-frontmatter>=1.1.0
